requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-js-asset==2.2.0
django-parler==2.3
django-redisboard==8.4.0
django-templated-mail==1.1.1
django-treebeard==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-dynamic-fields==0.4.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg==3.1.18
pycparser==2.21
PyJWT==2.8.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2023.4
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.2.0
